# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    CancelReleaseRequestRunsRequest:
      type: object
      required:
      - repoId
      - releaseRequestId
      properties:
        repoId:
          type: string
          description: The repository whose runs to stop
        releaseRequestId:
          type: string
          description: The release request whose work is withdrawn
    CancelReleaseRequestRunsResponse:
      type: object
      properties:
        runIds:
          type: array
          items:
            type: string
...
    TriggerEventRequest:
      type: object
      properties:
        name:
          type: string
          examples:
          - SoftwareRelease
          description: "The event signature, matched against a trigger file's `event:`"
        payload:
          type: object
          description: "The event payload, exactly as it would arrive off the bus"
        occurredAt:
          type: string
          examples:
          - 2026-08-04T09:00:00Z
          description: "ISO-8601 instant. Absent: now."
        eventId:
          type: string
          description: "UUID. Absent: a fresh random one — see the dedupe note."
    TriggerEventResult:
      type: object
      properties:
        eventId:
          type: string
        runIds:
          type: array
          items:
            type: string
        repositoriesRead:
          type: integer
          format: int32
        repositoriesSkipped:
          type: array
          items:
            type: string
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
paths:
...
  /ci/api/events/trigger:
    post:
      summary: Run every event pipeline that selects a caller-supplied domain event
      operationId: triggerEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerEventRequest"
        required: true
      responses:
        "200":
          description: "Evaluated, with the runs it recorded and which repositories\
            \ it could ask"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerEventResult"
        "400":
          description: "Blank name, missing payload, or an unparseable timestamp or\
            \ id"
        "403":
          description: The token covers no repository here — its project claim names
            a project this instance can place no repository in
        "503":
          description: Not evaluated — no candidate repository could be read. Retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerEventResult"
        "401":
          description: Not Authorized
      tags:
      - Ci Event Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
...
  /ci/api/runs/cancellations:
    post:
      summary: Cancel every unfinished run a repository has for one release request
      operationId: cancelReleaseRequestRuns
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelReleaseRequestRunsRequest"
        required: true
      responses:
        "202":
          description: The request's runs have been stopped or asked to stop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelReleaseRequestRunsResponse"
        "400":
          description: The repository id or the release request id is missing or invalid
        "403":
          description: "The token covers this repository for nobody — no project claim\
            \ and no platform-wide role, or a project claim this instance cannot place\
            \ the repository in"
        "401":
          description: Not Authorized
      tags:
      - Ci Run Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
...
info:
  title: qits-ci API
  version: 2026.1010.81424